Fixed databuffer not reading byte arrays properly

+++ Chip.Net/Chip.Net/Data/DataBuffer.cs
@@ -222,7 +222,7 @@ namespace Chip.Net.Data
 		/// </summary>
 		/// <returns></returns>
 		public byte[] ReadByteArray() {
-			var length = ReadByte();
+			var length = ReadUInt16();
 			byte[] ret = new byte[length];
 			for (int i = 0; i < length; i++)
 				ret[i] = ReadByte();
